Fix typo, Improve verbiage in raml file, and relocate test code
Some of the test code was put in shared locations.
For those blocks of code, move them into their specific test class.
