+++ ramls/items-lost-requiring-actual-costs.raml
@@ -1,12 +1,12 @@
 #%RAML 1.0
-title: Pick Slips
+title: Items Lost Requiring Actual Costs
 version: v0.2
 protocols: [ HTTP, HTTPS ]
 baseUri: http://localhost:9130
 
 documentation:
-  - title: API for fetching items requiring actual cost
-    content: <b>API for items requiring actual cost generation</b>
+  - title: API for fetching items lost requiring actual cost
+    content: <b>API for items lost requiring actual cost generation</b>
 
 types:
   loan: !include loan.json
