+++ src/test/java/api/support/APITests.java
@@ -99,7 +99,6 @@ public abstract class APITests {
   protected final ResourceClient instancesClient = ResourceClient.forInstances();
   protected final ResourceClient holdingsClient = ResourceClient.forHoldings();
   protected final ResourceClient itemsClient = ResourceClient.forItems();
-  protected final ResourceClient itemsLostRequiringActualCostsClient = ResourceClient.forItemsLostRequiringActualCosts();
 
   protected final ResourceClient loansClient = ResourceClient.forLoans();
   protected final ResourceClient accountsClient = ResourceClient.forAccounts();
